refactor: modularize StudentCourseDashboardView into separate tab components
