+++ src/views/StudentCourseDashboardView.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { ref, computed, watch, onMounted } from 'vue'
+import { ref, computed, watch } from 'vue'
 import { defineAsyncComponent } from 'vue'
 import { useRoute } from 'vue-router'
 import { useCoursesStore } from '@/stores/coursesStore'
@@ -7,9 +7,11 @@ import { useSectionsStore } from '@/stores/sectionsStore'
 import { useStudentsStore } from '@/stores/studentsStore'
 import { useQuizzesStore } from '@/stores/quizzesStore'
 import { useAuthStore } from '@/stores/authStore'
-import type { Student, StudentQuiz } from '@/interfaces/interfaces'
+import type { Student } from '@/interfaces/interfaces'
 const Header = defineAsyncComponent(() => import('@/components/Header.vue'))
-const StudentQuizList = defineAsyncComponent(() => import('@/components/student/StudentQuiz.vue'))
+const StudentCourseQuizzesTab = defineAsyncComponent(() => import('@/components/student/StudentCourseQuizzesTab.vue'))
+const StudentCourseScoreTab = defineAsyncComponent(() => import('@/components/student/StudentCourseScoreTab.vue'))
+const StudentCoursePeopleTab = defineAsyncComponent(() => import('@/components/student/StudentCoursePeopleTab.vue'))
 
 import bg1 from '@/assets/image/bg1.jpg'
 import bg2 from '@/assets/image/bg2.jpg'
@@ -19,15 +21,6 @@ import bg5 from '@/assets/image/bg5.jpg'
 
 // TYPES
 type TabKey = 'quizzes' | 'score' | 'people'
-
-interface MyScoreItem { 
-  title: string; 
-  score: number; 
-  total: number; 
-  percent: number; 
-  due: string; 
-  status: 'Answered' | 'Unanswered' 
-}
 
 // CONSTANTS
 const coverImages = [bg1, bg2, bg3, bg4, bg5]
@@ -43,8 +36,6 @@ const authStore = useAuthStore()
 
 // REF
 const activeTab = ref<TabKey>('quizzes')
-const quizViewMode = ref<'cards' | 'rows'>('rows')
-const scoreFilter = ref<'all' | 'answered' | 'unanswered'>('all')
 
 // COMPUTED
 const courseId = computed(() => Number(route.params.id || 0))
@@ -107,53 +98,6 @@ const students = computed<Student[]>(() => {
   })
 })
 
-const myCourseQuizzes = computed<StudentQuiz[]>(() => {
-  const all = quizzesStore.myStudentQuizzes
-  const courseName = currentCourse.value?.name || ''
-  return all.filter(q => q.subject === courseName)
-})
-
-const myScores = computed<MyScoreItem[]>(() => {
-  return myCourseQuizzes.value.map((q) => {
-    const history = quizzesStore.getQuizAttemptHistory(q.id)
-    const latestAttempt = history.length > 0 ? history[history.length - 1] : null
-
-    let total = 0
-    let score = 0
-
-    if (latestAttempt) {
-      score = latestAttempt.score
-      total = latestAttempt.totalPoints
-    } else {
-      const questions = quizzesStore.getStudentQuizQuestions(q.id)
-      total = questions.reduce((sum: number, qq: any) => sum + (typeof qq.points === 'number' ? qq.points : 0), 0)
-    }
-
-    const percent = total > 0 ? Math.round((score / total) * 100) : 0
-    const status: 'Answered' | 'Unanswered' = history.length > 0 ? 'Answered' : 'Unanswered'
-
-    return {
-      title: q.title,
-      score,
-      total,
-      percent,
-      due: q.dueDate,
-      status,
-    }
-  })
-})
-
-const filteredScores = computed<MyScoreItem[]>(() => {
-  if (scoreFilter.value === 'all') return myScores.value
-  if (scoreFilter.value === 'answered') return myScores.value.filter(s => s.status === 'Answered')
-  return myScores.value.filter(s => s.status === 'Unanswered')
-})
-
-
-onMounted(() => {
-	quizzesStore.loadAttemptHistoryFromStorage()
-})
-
 // WATCHERS
 watch(courseId, () => {
   activeTab.value = 'quizzes'
@@ -207,183 +151,21 @@ const getDeterministicIndex = (key: string) => {
         <button class="px-4 py-2 font-medium transition-colors cursor-pointer" :class="activeTab === 'people' ? 'text-blue-600 border-b-2 border-blue-600' : 'text-blue-400 hover:text-blue-600'" @click="activeTab = 'people'">People</button>
       </div>
 
-      <!-- Quizzes tab -->
-      <div v-show="activeTab === 'quizzes'" class="rounded-lg overflow-hidden">
-        <div class="px-6 flex justify-between items-center">
-          <h2 class="text-xl font-semibold text-blue-800">My Quizzes</h2>
-          <div v-if="myCourseQuizzes.length > 0" class="inline-flex items-center gap-2 bg-gray-100 p-1 rounded-lg">
-            <button
-              class="relative px-3 py-2 rounded-md text-sm font-medium transition-all duration-200 ease-in-out"
-              :class="quizViewMode === 'cards' ? 'bg-white text-blue-600 shadow-sm' : 'text-gray-600 hover:text-gray-900'"
-              @click="quizViewMode = 'cards'"
-              aria-label="Cards view"
-              title="Cards view"
-            >
-              <i class="fas fa-grip"></i>
-            </button>
-            <button
-              class="relative px-3 py-2 rounded-md text-sm font-medium transition-all duration-200 ease-in-out"
-              :class="quizViewMode === 'rows' ? 'bg-white text-blue-600 shadow-sm' : 'text-gray-600 hover:text-gray-900'"
-              @click="quizViewMode = 'rows'"
-              aria-label="Rows view"
-              title="Rows view"
-            >
-              <i class="fas fa-list"></i>
-            </button>
-          </div>
-        </div>
-        
-        <!-- Empty State -->
-        <div v-if="myCourseQuizzes.length === 0" class="p-12 flex flex-col items-center justify-center text-center">
-          <div class="relative mb-6">
-            <div class="w-24 h-24 bg-gradient-to-br from-blue-50 to-indigo-100 rounded-full flex items-center justify-center">
-              <i class="fas fa-clipboard-list text-4xl text-blue-400"></i>
-            </div>
-            <div class="absolute -bottom-1 -right-1 w-8 h-8 bg-blue-500 rounded-full flex items-center justify-center shadow-lg">
-              <i class="fas fa-clock text-white text-sm"></i>
-            </div>
-          </div>
-          <h3 class="text-xl font-semibold text-gray-800 mb-2">No Quizzes Available Yet</h3>
-          <p class="text-gray-500 max-w-md mb-6">
-            There are currently no quizzes assigned for this course. Check back later or contact your instructor for more information.
-          </p>
-          <div class="flex items-center gap-2 text-sm text-gray-400">
-            <i class="fas fa-info-circle"></i>
-            <span>New quizzes will appear here when they're assigned</span>
-          </div>
-        </div>
+      <StudentCourseQuizzesTab
+        v-show="activeTab === 'quizzes'"
+        :courseName="currentCourse?.name || ''"
+      />
 
-        <!-- Quiz List -->
-        <div v-else class="p-6">
-          <StudentQuizList :quizzes="myCourseQuizzes" :hideHeader="true" :viewMode="quizViewMode" />
-        </div>
-      </div>
+      <StudentCourseScoreTab
+        v-show="activeTab === 'score'"
+        :courseName="currentCourse?.name || ''"
+      />
 
-      <!-- Score tab  -->
-      <div v-show="activeTab === 'score'" class="space-y-6">
-        <!-- No Scores at All -->
-        <div v-if="myScores.length === 0" class="bg-white rounded-lg shadow border border-gray-200 p-12 flex flex-col items-center justify-center text-center">
-          <div class="relative mb-6">
-            <div class="w-24 h-24 bg-gradient-to-br from-blue-50 to-indigo-100 rounded-full flex items-center justify-center">
-              <i class="fas fa-chart-line text-4xl text-blue-400"></i>
-            </div>
-            <div class="absolute -bottom-1 -right-1 w-8 h-8 bg-blue-500 rounded-full flex items-center justify-center shadow-lg">
-              <i class="fas fa-star text-white text-sm"></i>
-            </div>
-          </div>
-          <h3 class="text-xl font-semibold text-gray-800 mb-2">No Scores Yet</h3>
-          <p class="text-gray-500 max-w-md mb-6">
-            You haven't completed any quizzes for this course yet. Start taking quizzes to see your scores here!
-          </p>
-          <div class="flex items-center gap-2 text-sm text-gray-400">
-            <i class="fas fa-info-circle"></i>
-            <span>Your quiz scores and progress will be tracked here</span>
-          </div>
-        </div>
-
-        <!-- Scores Available -->
-        <div v-else class="bg-white rounded-lg shadow border border-gray-200">
-          <!-- Filter dropdown -->
-          <div class="px-6 py-4 border-b border-gray-200">
-            <div class="relative w-full max-w-sm">
-              <label class="block text-xs text-gray-500 mb-2">Task filter</label>
-              <select 
-                v-model="scoreFilter" 
-                class="w-full px-4 py-2 border border-gray-300 rounded-md bg-white text-gray-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-blue-500 appearance-none cursor-pointer"
-              >
-                <option value="all">All</option>
-                <option value="answered">Answered</option>
-                <option value="unanswered">Unanswered</option>
-              </select>
-              <i class="fas fa-chevron-down absolute right-4 top-10 text-gray-400 pointer-events-none"></i>
-            </div>
-          </div>
-
-          <!-- Scores list -->
-          <div class="divide-y divide-gray-100">
-            <div 
-              v-for="q in filteredScores" 
-              :key="q.title" 
-              class="px-6 py-5 hover:bg-gray-50 transition-colors cursor-pointer flex items-center justify-between"
-            >
-              <div class="flex-1">
-                <h4 class="font-medium text-gray-900 mb-1">{{ q.title }}</h4>
-                <p class="text-sm text-gray-500">Due {{ q.due }}</p>
-              </div>
-              <div class="flex items-center gap-6">
-                <div class="text-right">
-                  <div class="text-sm text-gray-500 mb-1">Status</div>
-                  <div class="text-sm font-medium" :class="q.status === 'Answered' ? 'text-gray-700' : 'text-red-600'">
-                    {{ q.status }}
-                  </div>
-                </div>
-                <div class="text-right min-w-[80px]">
-                  <div class="text-2xl font-semibold text-gray-900">{{ q.score }}/{{ q.total }}</div>
-                </div>
-              </div>
-            </div>
-            
-            <!-- Empty Filtered Results -->
-            <div v-if="filteredScores.length === 0" class="px-6 py-12 flex flex-col items-center justify-center text-center">
-              <div class="relative mb-6">
-                <div class="w-20 h-20 bg-gradient-to-br from-gray-50 to-gray-100 rounded-full flex items-center justify-center">
-                  <i class="fas fa-filter text-3xl text-gray-300"></i>
-                </div>
-              </div>
-              <h4 class="text-lg font-semibold text-gray-700 mb-2">No Matching Scores</h4>
-              <p class="text-gray-500 max-w-sm mb-4">
-                No scores match your current filter. Try selecting a different filter option.
-              </p>
-              <button 
-                @click="scoreFilter = 'all'" 
-                class="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-700 transition-colors text-sm font-medium"
-              >
-                <i class="fas fa-undo mr-2"></i>Clear Filter
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <!-- People tab -->
-      <div v-show="activeTab === 'people'" class="space-y-8">
-        <!-- Teachers Section -->
-        <div class="bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden">
-          <div class="px-6 py-4 border-b border-gray-100">
-            <h2 class="text-2xl font-semibold text-gray-900">Teachers</h2>
-          </div>
-          <div class="px-6 py-5">
-            <div class="flex items-center gap-4 hover:bg-gray-50 -mx-6 px-6 py-3 transition-colors">
-              <div class="h-10 w-10 rounded-full bg-indigo-600 flex items-center justify-center text-white font-bold flex-shrink-0">
-                {{ (currentCourse?.teacher || 'U').split(' ').map(n => n[0]).slice(0,2).join('').toUpperCase() }}
-              </div>
-              <div class="font-medium text-gray-900">{{ currentCourse?.teacher }}</div>
-            </div>
-          </div>
-        </div>
-
-        <!-- Classmates Section -->
-        <div class="bg-white rounded-lg shadow-sm border border-gray-200 overflow-hidden">
-          <div class="px-6 py-4 border-b border-gray-100 flex items-center justify-between">
-            <h2 class="text-2xl font-semibold text-gray-900">Classmates</h2>
-            <span class="text-sm text-gray-500">{{ students.length }} students</span>
-          </div>
-          <div class="divide-y divide-gray-100">
-            <div 
-              v-for="s in students" 
-              :key="s.id" 
-              class="px-6 py-4 flex items-center gap-4 hover:bg-gray-50 transition-colors"
-            >
-              <img 
-                :src="s.avatar" 
-                :alt="s.name" 
-                class="h-10 w-10 rounded-full object-cover flex-shrink-0" 
-              />
-              <div class="font-medium text-gray-900">{{ s.name }}</div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <StudentCoursePeopleTab
+        v-show="activeTab === 'people'"
+        :teacherName="currentCourse?.teacher || null"
+        :students="students"
+      />
     </div>
   </div>
 </template>
@@ -391,19 +173,6 @@ const getDeterministicIndex = (key: string) => {
 <style scoped>
 .classroom-banner {
   height: 250px;
-}
-
-.grade-progress {
-  height: 8px;
-  border-radius: 4px;
-  background-color: #e0e7ff;
-}
-
-.grade-progress-fill {
-  height: 100%;
-  border-radius: 4px;
-  background: linear-gradient(90deg, #3b82f6, #1d4ed8);
-  transition: width 0.5s ease;
 }
 
 @media (max-width: 768px) {
